+++ DTCD-Visualization-Gauge/src/PluginComponent.vue
@@ -12,7 +12,9 @@
 <script>
 export default {
   name: 'PluginComponent',
-  data: () => ({
+  data: ({ $root }) => ({
+    notificationSystem: $root.notificationSystem,
+    guid: $root.guid,
     /** Gauge technical data. */
     svg: null,
     radius: 0,
@@ -28,6 +30,10 @@ export default {
     colValue: 'value',
     segments: [],
     dataset: [],
+    panelSize: {
+      height:200,
+      width:200
+    },
   }),
   computed: {
     computedTitle() {
@@ -43,14 +49,25 @@ export default {
   },
   mounted() {
     const { svgContainer } = this.$refs;
+    const rect = svgContainer.getBoundingClientRect()
+    this.panelSize =  {
+      width: rect.width,
+      height: rect.height,
+    }
     const attrs = svgContainer.getAttributeNames();
     this.dataAttr = attrs.find(attr => attr.startsWith('data-'));
   },
-  methods: {
-    setConfigProp(prop, value) {
-      this[prop] = value;
+  watch: {
+    panelSize: {
+      deep: true,
+      handler(val, old) {
+        if (JSON.stringify(val) !== JSON.stringify(old)) {
+          this.render()
+        }
+      },
     },
-
+  },
+  methods: {
     setTitle(text = '') {
       this.title = text;
       this.render();
@@ -70,7 +87,9 @@ export default {
       this.segments = segments;
       this.render();
     },
-
+    setPanelSize(panelSize) {
+      this.panelSize = panelSize
+    },
     setDataset(data = []) {
       this.dataset = data;
       this.render();
@@ -95,6 +114,19 @@ export default {
         const segment = this.segments.find(s => s.range[1] === max);
         this.valueColor = segment.color;
       }
+
+      const floatValue = Number.parseFloat(this.value);
+      if (!isNaN(floatValue) && (floatValue < min || floatValue > max)) {
+        this.$root.createNotification(
+          `Gauge: ${this.guid}`,
+          'Значение вышло за пределы шкалы',
+          {
+            floatMode: true,
+            tag: `${this.guid}-outside-value`,
+            type: 'warning',
+          }
+        )
+      }
     },
 
     render() {
@@ -115,9 +147,9 @@ export default {
     prepareRenderData() {
       const { mainContainer, svgContainer } = this.$refs;
       const sizeCutting = this.computedTitle.length <= 0 ? 70 : 50;
-      const isContainerSizesEqual = mainContainer.offsetWidth === mainContainer.offsetHeight;
 
-      let { offsetWidth: width, offsetHeight: height } = svgContainer;
+      let { width, height } = this.panelSize;
+      const isContainerSizesEqual = width === height;
 
       if (isContainerSizesEqual) {
         width -= sizeCutting;
@@ -126,7 +158,6 @@ export default {
 
       const translateWidth = isContainerSizesEqual ? width + sizeCutting : width;
       const translateHeight = isContainerSizesEqual ? height + sizeCutting : height;
-
       this.svg = d3
         .select(svgContainer)
         .append('svg')
@@ -141,7 +172,6 @@ export default {
       this.valueRange = d3.extent(this.segments.map(s => s.range).flat());
 
       this.radius = Math.min(width, height) / 2;
-      this.radius += width > height ? 25 : -25;
 
       this.arrowLength = this.radius - this.segmentWidth / 2;
       this.scale = d3.scaleLinear().range(angles).domain(this.valueRange);
@@ -206,7 +236,12 @@ export default {
         .innerRadius(this.radius - this.segmentWidth)
         .startAngle(this.scale(start))
         .endAngle(this.scale(end));
-      this.svg.append('path').attr('fill', color).attr('d', arc);
+      this.svg.append('path')
+        .attr('fill', color)
+        .attr('d', arc)
+        .on('click', () => {
+          this.$root.publishEventClicked({ range, color })
+        });
     },
 
     addTextElement({ x, y, text, className }) {
@@ -232,7 +267,7 @@ export default {
   height: 100%
   display: flex
   flex-direction: column
-  font-family: 'Proxima Nova'
+  font-family: 'Proxima Nova', serif
 
   .NoData
     flex-grow: 1